Add multiple components and services to main app layout
Added slacker service as a Service in Service that communicates with the server.
Added message component and service as prototype for user response.
Added slacker-detail as a child to slacker for individual slacker descriptions.
Added dashboard for easy navigation around the web app.

Need to accommodate HTTP request methods to retrieve data from server.